support 6.1 api update stickers and premium stickers
The maximum value of the field file_size in the classes Animation, Audio, Document, Video, Voice, and File can no longer be stored in a signed 32-bit integer type. This change is necessary to support 4GB files uploaded by premium accounts.
Added the field is_premium to the class User.
Added the field premium_animation to the class Sticker.